Move WithExit to runcopts package
